Dev 2.2 (#196)
* Simplify IR and begin introducing meta-symbols

Simplify and clean up the structure of IR, which now has only one notion of symbol.
One special case is meta-symbols: new symbols introduced when parsing the body of functions, such as a block or an if statements.
Meta symbols also have an unique qualifier, which is determined based on the nature of the symbol (e.g. the else branch of a conditional) or the position in the body.

* IR: unify class body and function body.

* IR: dump symbol body.

* IR block: represent statements directly in the symbol's body.

* IR: don't create a separate block metasymbol for function bodies.

Instead, use the symbol's `body` field.

* rename

* IR: add parent node w.r.t. control flow.

* IR: use array notation for the statement index in the scope.

* IR: remove block symbol.

Blocks are inlined in the symbols that own them.

* IR: add test with nested conditionals.

* IR: call and string.

* IR: define Expression separately from Statement.

* IR: remove statement and use Item.

Split up several symbols into several items in the case of mutual definitions: let a = ... and b = ... instead of putting several symbols in one statement. This gives an accurate number of statements in the body.

* IR: rename if branch to block.

* IR: begin using a counter for symbols.

* IR: use counter to name metasymbols.

* IR: remove "body" and indexing "[...]" from metasymbols.

* IR: expression statements.

* IR: parse expressions as code and replace symbols with their name.

* IR: correct body update in update_dummy_symbol.

* IR: make case conditions expressions.

* IR: expressions are fully substituted string.

* IR: include metasymbols in project map.

* IR: append symbols to parent's body during symbol add

This should ensure a consistent setting of body symbols.

* IR keep a counter per name.

* IR: show details of metastymbols in project map.

* IR opt: don't create an expression statement symbol if the expression itself needs a symbol.

* IR: walk assignment and binary operator expressions

* format

* IR: add "guard" nodes in conditionals.

* IR: use a body node for branches in conditionals.

* IR: guard and body nodes have their own counters.